+++ laktory/models/datasinks/__init__.py
@@ -1,4 +1,5 @@
 from .basedatasink import BaseDataSink
+from .customwriter import CustomWriter
 from .filedatasink import FileDataSink
 from .hivemetastoredatasink import HiveMetastoreDataSink
 from .mergecdcoptions import DataSinkMergeCDCOptions